+++ protocol-designer/src/steplist/actions/index.js
@@ -1,4 +1,3 @@
 // @flow
 export * from './actions'
 export * from './types'
-export * from './thunks'
