+++ protocol-designer/src/ui/steps/constants.js
@@ -1,0 +1,2 @@
+// @flow
+export const MAIN_CONTENT_FORCED_SCROLL_CLASSNAME = 'main_content_forced_scroll'
