+++ protocol-designer/src/ui/steps/index.js
@@ -1,7 +1,10 @@
 // @flow
 
-import * as actions from './actions'
-
+import * as _actions from './actions/actions'
+import * as _thunks from './actions/thunks'
 export { default as rootReducer } from './reducers'
 export { default as selectors } from './selectors'
-export { actions }
+export * from './constants'
+export type * from './actions/types'
+
+export const actions = { ..._actions, ..._thunks }
